+++ pyeudiw/sd_jwt/utils/demo_utils.py
@@ -1,9 +1,6 @@
-import base64
 import logging
 import random
 import sys
-from typing import Union
-
 import yaml
 from cryptojwt.jwk.jwk import key_from_jwk_dict
 
@@ -28,25 +25,6 @@ def load_yaml_settings(file):
         ]
 
     return settings
-
-
-def print_repr(values: Union[str, list], nlines=2):
-    value = "\n".join(values) if isinstance(values, (list, tuple)) else values
-    _nlines = "\n" * nlines if nlines else ""
-    print(value, end=_nlines)
-
-
-def print_decoded_repr(value: str, nlines=2):
-    seq = []
-    for i in value.split("."):
-        try:
-            padded = f"{i}{'=' * divmod(len(i),4)[1]}"
-            seq.append(f"{base64.urlsafe_b64decode(padded).decode()}")
-        except Exception as e:
-            logging.debug(f"{e} - for value: {i}")
-            seq.append(i)
-    _nlines = "\n" * nlines if nlines else ""
-    print("\n.\n".join(seq), end=_nlines)
 
 
 def get_jwk(jwk_kwargs: dict = {}, no_randomness: bool = False, random_seed: int = 0):
